// src/app/Models/project.ts
export class Project {

    ProjectId: number;
    ProjectName: string
    ProjectPriority: number;
    DateReqd: string;
    StartDate: Date;
    EndDate: Date;
    ManagerId: number;
    ProjectStatus: string;
    AddDate: Date;
    UpdtDate: Date;

    ManagerName: string;
    TaskCount: number;
    CompletedTasks: number;

}